perf(fp-frame): speed Kendall inversion count 1.52x
br-frankenpandas-sad5y

Replace the profiled recursive usize merge inversion counter with a bounded-rank Fenwick counter over y_rank_by_row[x_order].

Profile target: df_kendall 512 20, before profile had count_usize_inversions_recursive at about 69.6% self samples.

Hyperfine via RCH release-perf perf_profile: before 62.4 ms +/- 6.5 ms; after 41.0 ms +/- 1.8 ms; speedup 1.5213x; mean reduction 34.26%.

Behavior proof: Kendall ordering, tie fallback, integer discordant-pair arithmetic, floating formula, and RNG inputs unchanged; golden sha256 stayed dcce8a4e1f13d887361650ec24fcfee13c6c4c3954936fd124a1feb53cf125ce.

Verification: rch cargo test -p fp-frame --lib kendall; rch cargo check -p fp-frame --all-targets; rch cargo clippy -p fp-frame --all-targets -- -D warnings.

+++ crates/fp-frame/examples/bench_kendall.rs
@@ -90,7 +90,7 @@ fn cross_check() -> (usize, usize) {
     let (mut ok, mut bad) = (0usize, 0usize);
     for _ in 0..4000 {
         let n = (next() % 30) as usize + 2;
-        let range = (next() % 6) as u64 + 1; // small range -> many ties
+        let range = next() % 6 + 1; // small range -> many ties
         let xs: Vec<f64> = (0..n).map(|_| (next() % range) as f64).collect();
         let ys: Vec<f64> = (0..n).map(|_| (next() % range) as f64).collect();
         let got = s_from(&xs).corr_kendall(&s_from(&ys)).unwrap();
